#52 broadcast load info

+++ src/main/java/pl/edu/agh/hiputs/communication/model/messages/LoadInfoMessage.java
@@ -6,7 +6,7 @@ import pl.edu.agh.hiputs.communication.model.MessagesTypeEnum;
 @Value
 public class LoadInfoMessage implements Message {
 
-  int carCost;
+  long carCost;
   long time;
   String mapFragmentId;
